Remove mock authorization service
