--- src/main/java/edu/kpi/testcourse/auth/AuthorizationMockService.java
@@ -1,10 +1,0 @@
-package edu.kpi.testcourse.auth;
-
-import edu.kpi.testcourse.exception.auth.UnauthorizedException;
-
-/**
- * Mock service to provide user authorization by completely random JWT token.
- */
-public interface AuthorizationMockService {
-  String authorizeUser(String jwtToken) throws UnauthorizedException;
-}
